Refactor: SCSS 변수 적용 및 스타일 구조 개선
