+++ package.json
@@ -24,6 +24,7 @@
     "eslint-plugin-react-hooks": "^5.1.0",
     "eslint-plugin-react-refresh": "^0.4.19",
     "globals": "^15.15.0",
+    "sass": "^1.86.3",
     "vite": "^6.2.0"
   }
 }
