+++ src/App.jsx
@@ -7,7 +7,7 @@ import Items from "./pages/Items";
 import FAQ from "./pages/FAQ";
 import Privacy from "./pages/Privacy";
 import "./styles/reset.css";
-import "./styles/variable.css";
+import "./styles/utils/variable.css";
 import "./styles/global.css";
 
 function App() {
