+++ src/components/home/HomeBottomSection.jsx
@@ -1,5 +1,5 @@
-import common from "../../styles/components/HomeCommonSection.module.css";
-import styles from "../../styles/components/HomeBottomSection.module.css";
+import common from "../../styles/components/home/HomeCommonSection.module.scss";
+import styles from "../../styles/components/home/HomeBottomSection.module.scss";
 import bottomImg from "../../assets/images/Img_home_bottom.png";
 import classNames from "classnames";
