+++ src/components/home/HomeTopSection.jsx
@@ -1,6 +1,6 @@
 import classNames from "classnames";
-import styles from "../../styles/components/HomeTopSection.module.css";
-import common from "../../styles/components/HomeCommonSection.module.css";
+import styles from "../../styles/components/home/HomeTopSection.module.scss";
+import common from "../../styles/components/home/HomeCommonSection.module.scss";
 import topImage from "../../assets/images/Img_home_top.png";
 import { Link } from "react-router-dom";
