--- src/styles/components/HomeBottomSection.module.css
@@ -1,36 +1,0 @@
-.bottomSection {
-  gap: 69px;
-}
-
-.bottom {
-  margin: 112.5px 0 112.5px;
-}
-
-.bottomImg {
-  width: 746px;
-  height: 397px;
-}
-
-@media (max-width: 1199px) and (min-width: 768px) {
-  .bottomSection {
-    height: 927px;
-  }
-
-  .bottom {
-    width: 295px;
-    height: 112px;
-    margin: 201px 0 0;
-  }
-
-  .bottomImg {
-    width: 746px;
-    height: 397px;
-  }
-}
-
-@media (max-width: 767px) {
-  .bottomImg {
-    width: 375px;
-    height: 198px;
-  }
-}
