--- src/styles/components/HomeCommonSection.module.css
@@ -1,70 +1,0 @@
-.background {
-  background-color: var(--skyblue);
-  display: flex;
-  justify-content: center;
-  align-items: end;
-  height: 540px;
-}
-
-.bgDescription {
-  justify-content: center;
-  align-items: center;
-  padding-bottom: 60px;
-}
-
-.bgDescription h2 {
-  margin: 0 0 32px 0;
-}
-
-/* .lineBreak {
-  display: inline;
-} */
-
-h2 {
-  font-size: 40px;
-  font-weight: 700;
-  line-height: 56px;
-  color: var(--gray700);
-}
-
-@media (min-width: 1200px) {
-  .lineBreak {
-    display: block;
-  }
-}
-
-@media (max-width: 1199px) and (min-width: 768px) {
-  .background {
-    flex-direction: column;
-    justify-content: space-between;
-    align-items: center;
-    height: 771px;
-  }
-
-  h2 {
-    font-weight: 700;
-    font-size: 40px;
-    line-height: 140%;
-    letter-spacing: 0%;
-    text-align: center;
-    vertical-align: middle;
-  }
-}
-
-@media (max-width: 767px) {
-  .background {
-    flex-direction: column;
-    justify-content: space-between;
-    align-items: center;
-    height: auto;
-  }
-
-  h2 {
-    font-size: 32px;
-    font-weight: 700;
-    line-height: 140%;
-    letter-spacing: 0%;
-    text-align: center;
-    vertical-align: middle;
-  }
-}
