--- src/styles/components/HomeFooter.module.css
@@ -1,112 +1,0 @@
-footer {
-  background-color: var(--gray900);
-  justify-content: center;
-  height: 160px;
-}
-
-.footerContainer {
-  color: var(--gray400);
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  padding: 32px;
-  margin: 0 auto;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 100%;
-  letter-spacing: 0%;
-  text-align: center;
-}
-
-.footerContainer h5 {
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 100%;
-  letter-spacing: 0%;
-  text-align: center;
-  grid-area: h;
-}
-
-.privacyFAQ {
-  display: flex;
-  gap: 30px;
-  grid-area: p;
-}
-
-.policyLinks {
-  text-decoration: none;
-  color: var(--gray200);
-}
-
-.icons {
-  display: flex;
-  gap: 12px;
-  padding: 0;
-  list-style: none;
-}
-
-.icons img {
-  width: 20px;
-  height: 20px;
-}
-
-@media (max-width: 1199px) and (min-width: 768px) {
-  .footerContainer {
-    width: auto;
-    padding: 32px 0 108px;
-    margin: 0 104px;
-    font-family: Abel;
-  }
-
-  .privacyFAQ {
-    display: flex;
-    gap: 30px;
-  }
-
-  .policyLinks {
-    text-decoration: none;
-    color: var(--gray200);
-  }
-}
-
-@media (max-width: 767px) {
-  .footerContainer {
-    justify-content: space-between;
-    align-items: center;
-    text-align: center;
-    height: auto;
-    padding: 32px;
-    font-family: Arial;
-    position: relative;
-  }
-
-  .footerContainer h5 {
-    display: block;
-    text-align: center;
-    position: absolute;
-    top: 110px;
-    left: 32px;
-  }
-
-  .privacyFAQ {
-    display: flex;
-    gap: 30px;
-  }
-
-  .policyLinks {
-    text-decoration: none;
-    color: var(--gray200);
-  }
-
-  .icons {
-    display: flex;
-    gap: 12px;
-    padding: 0;
-  }
-
-  .icons img {
-    width: 18px;
-    height: 18px;
-  }
-}
