+++ src/styles/utils/variable.css
@@ -8,9 +8,11 @@
   --gray200: #e5e7eb;
   --gray100: #f3f4f6;
   --gray50: #f9fafb;
+
   --blue100: #3692ff;
   --blue200: #1967d6;
   --blue300: #1251aa;
+
   --red: #f74747;
   --white: #ffffff;
   --background: #e6f2ff;
